requirements.txt
asgiref==3.5.2
boto3==1.24.44
botocore==1.27.44
dj-database-url==1.0.0
Django==4.0.6
django-cors-headers==3.13.0
django-extensions==3.2.0
django-on-heroku==1.1.2
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
whitenoise==6.2.0
